- dialog event 처리 추가 - event 간 view show/hide 처리

+++ app/src/main/java/com/example/clone_daum/ui/search/SearchViewModel.kt
@@ -1,8 +1,11 @@
 package com.example.clone_daum.ui.search
 
 import android.app.Application
+import android.content.SharedPreferences
+import android.view.View
 import androidx.annotation.StringRes
 import androidx.databinding.ObservableField
+import androidx.databinding.ObservableInt
 import com.example.clone_daum.R
 import com.example.clone_daum.model.DbRepository
 import com.example.clone_daum.model.local.SearchKeyword
@@ -24,22 +27,37 @@ class SearchViewModel @Inject constructor(
 ) : RecyclerViewModel<SearchKeyword>(app) {
     companion object {
         private val mLog = LoggerFactory.getLogger(SearchViewModel::class.java)
+
+        const val RECENT_SEARCH_LIMIT = 4L
+        const val K_RECENT_SEARCH = "search-recent-search"
     }
 
-    val searchKeyword = ObservableField<String>()
+    val searchKeyword            = ObservableField<String>()
+    var showSearchRecyclerLayout = prefs().getBoolean(K_RECENT_SEARCH, true)
+    val toggleRecentSearchText   = ObservableInt()
+    val toggleEmptyAreaText      = ObservableInt()
 
-    val searchEvent   = SingleLiveEvent<String>()
-    val closeEvent    = SingleLiveEvent<Void>()
-    val errorEvent    = SingleLiveEvent<String>()
+    val visibleSearchRecycler = ObservableInt(View.VISIBLE)
+    val visibleSearchEmpty    = ObservableInt(View.GONE)
+
+    val searchEvent      = SingleLiveEvent<String>()
+    val closeEvent       = SingleLiveEvent<Void>()
+    val dlgEvent         = SingleLiveEvent<DialogParam>()
+    val errorEvent       = SingleLiveEvent<String>()
 
     fun init() {
         initAdapter("search_recycler_history_item")
-        items.set(db.searchHistoryDao.search().limit(4).blockingFirst())
+
+        val searchList = db.searchHistoryDao.search().limit(RECENT_SEARCH_LIMIT).blockingFirst()
+        items.set(searchList)
+        visibleSearchRecycler(searchList.size > 0)
     }
 
     fun reloadData() {
-        disposable.add(db.searchHistoryDao.search().limit(4).subscribe {
+        disposable.add(db.searchHistoryDao.search().limit(RECENT_SEARCH_LIMIT).subscribe {
             items.set(it)
+
+            visibleSearchRecycler(it.size > 0)
         })
     }
 
@@ -63,17 +81,48 @@ class SearchViewModel @Inject constructor(
         } ?: errorEvent(R.string.error_empty_keyword)
     }
 
-    private fun errorEvent(@StringRes resid: Int) =
-        errorEvent(string(resid))
-
-    private fun errorEvent(msg: String?) {
-        mLog.error("ERROR: $msg")
-
-        msg?.run { errorEvent.value = this }
+    fun toggleRecentSearch() {
+        if (showSearchRecyclerLayout) {
+            dlgEvent.value = DialogParam(
+                title    = string(R.string.dlg_title_stop_using_recent_search),
+                message  = string(R.string.dlg_msg_do_you_want_stop_using_recent_search),
+                listener = { res, _ -> if (res) toggleSearchRecyclerLayout() },
+                positiveStr = string(android.R.string.ok),
+                negativeStr = string(android.R.string.cancel)
+            )
+        } else {
+            toggleSearchRecyclerLayout()
+        }
     }
 
-    fun closeSearchHistory() {
-        closeEvent.call()
+    private fun toggleSearchRecyclerLayout() {
+        showSearchRecyclerLayout = !showSearchRecyclerLayout
+
+        prefs().edit { putBoolean(K_RECENT_SEARCH, showSearchRecyclerLayout) }
+
+        visibleSearchRecycler(items.get()!!.size > 0)
+    }
+
+    private fun visibleSearchRecycler(res: Boolean) {
+        if (mLog.isDebugEnabled) {
+            mLog.debug("EXIST LIST $res, SHOW LAYOUT $showSearchRecyclerLayout")
+        }
+
+        if (res && showSearchRecyclerLayout) {
+            visibleSearchRecycler.set(View.VISIBLE)
+            visibleSearchEmpty.set(View.GONE)
+        } else {
+            visibleSearchRecycler.set(View.GONE)
+            visibleSearchEmpty.set(View.VISIBLE)
+        }
+
+        if (showSearchRecyclerLayout) {
+            toggleRecentSearchText.set(R.string.search_turn_off_recent_search)
+            toggleEmptyAreaText.set(R.string.search_empty_recent_search)
+        } else {
+            toggleRecentSearchText.set(R.string.search_turn_on_recent_search)
+            toggleEmptyAreaText.set(R.string.search_off_recent_search)
+        }
     }
 
     fun deleteHistory(item: SearchKeyword) {
@@ -89,11 +138,21 @@ class SearchViewModel @Inject constructor(
     }
 
     fun deleteAllHistory() {
-        // delete query 는 왜? Completable 이 안되는가?
-        ioThread {
-            db.searchHistoryDao.deleteAll()
-            reloadData()
-        }
+        dlgEvent.value = DialogParam(
+            title    = string(R.string.dlg_title_delete_all_searched_list),
+            message  = string(R.string.dlg_msg_delete_all_searched_list),
+            listener = { res, _ ->
+                if (res) {
+                    // delete query 는 왜? Completable 이 안되는가?
+                    ioThread {
+                        db.searchHistoryDao.deleteAll()
+                        reloadData()
+                    }
+                }
+            },
+            positiveStr = string(android.R.string.ok),
+            negativeStr = string(android.R.string.cancel)
+        )
     }
 
     fun searchCancel() {
@@ -101,4 +160,19 @@ class SearchViewModel @Inject constructor(
     }
 
     fun dateConvert(date: Long) = date.toDate("MM.dd.")
+
+    ////////////////////////////////////////////////////////////////////////////////////
+    //
+    //
+    //
+    ////////////////////////////////////////////////////////////////////////////////////
+
+    private fun errorEvent(@StringRes resid: Int) =
+        errorEvent(string(resid))
+
+    private fun errorEvent(msg: String?) {
+        mLog.error("ERROR: $msg")
+
+        msg?.run { errorEvent.value = this }
+    }
 }
